phish-mllm-benchmark/phish-mllm-benchmark/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phish-mllm-benchmark"
version = "0.1.0"
description = "Phishing detection benchmark with open MLLMs (Qwen2-VL, LLaVA-OneVision, Phi-3.5)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Your Name"}]
dependencies = [
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "tqdm>=4.66.0",
    "json5>=0.9.14",
]

[project.scripts]
phishbench = "phishbench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
